Character class improved (and fully documented)
Now, comparisons using the ">", "<", ">=", "<=" and "==" operators can take place without any errors. These comparisons are employed throughout the class itself, which is now able to show the next_characters list in an ordered fashion.

Additionally, the documentation is complete to the functions present at the moment. More details can be provided (and probably will); however, at this point I can say that at least every function has a minimal documentation clarifying aspects such as the parameters used (required or not) and the output.

Finally, minor changes are here as well, such as the uniformization of the line endings (I don't know why, but Visual Studio all of a sudden removed my extensions. I think it is something related with VS 2019, but I'll have to verify that later).