.__repr__() method added and bug fixes
The CharacterTree object representation was added, omitting things that were useful just with Characters.

I just stopped to think and realized I could, with a little more intricacy, merge both classes! This would be especially interesting when the fact that subtrees are trees as well is considered. Anyway, I'll think about the possibility of, down the road, merge them both.