.lower() part of self.character assignment line commented
This line was in the middle of the way of the from_csv() implementation seen under __init__.py, since it allowed the possibility of considering the case of the letters when adding them to the CharacterTree object to be created. The problems have diminished, but not vanished entirely; therefore, I'll have to do further analysis in order to grasp the problems (which I suspect are near the __eq__() method) and mitigate them.