.insert() apparently works!
I tried a couple of cases, such as "mamão", "maniçoba" (ma-), "menino" (m-) and it worked properly, changing the radixes to False and creating new ones to the remaining of the word. However, when I tried to insert "menina", it inserted the whole word in the root node (despite the existance of the "m" radix)! Therefore, some polishing is needed.