RadixTree.insert() significantly improved
The previously extremely unreliable .insert() method got quite some updates, and it seems to perform way better now, with combos such as "maniçoba", "mamão", "menino" and "menina" being inserted and qualified as final without any issues. However, additional testing is required with combos such as "abacate" and "aba", for instance (where the second word is completely present in the Tree, but not marked as final.