.__contains__() and .__repr__() implemented, .insert() improved
The .__contains__() method now works, allowing us to see if a given word is in the Tree object or not. In addition, the .__repr__() method was implemented to properly show the necessary information, including the number of loaded words (which, so far, seems to match the one of the CharacterTree object). Finally, the .insert() method seems to be complete and stable, and got its debugging print statements commented out to make the loading process faster during testing. I'll probably add a verbose function to them later on; so far, however, let's leave them at that.