.insert() fixed, .remove() created, comments deleted
There was a problem with the .insert() method which prevented words that were contained in words added to the RadixTree from being inserted, i.e. trying to insert "aba" when "abacate" was inserted.

In addition, the .remove() method was created, not only removing the words (at least by setting Radix.is_final to False), but also attempting to optimize the RadixTree by merging radices with only a single child and removing non-final Radix objects with no children (len(Radix.next_radices)==0). So far, so good.