misc.py and its references removed
Since the get_radices() function ended up not being used at all due to a change of strategy in the implementation of the RadixTree methods, it was removed, along with misc.py (which only contained this function). The references to this file, present under __init__.py and RadixTree.py, were also removed.